chore(clayui.com): Extract useStickyState into it's own file, and optimize it, show Expand text and hide Copy button when code is collapsed
